Move registration of script commands to main.cpp (this isn't platform agnostic)^

+++ tgbdual_desktop/src/gameboy.cpp
@@ -8,7 +8,7 @@ gameboy::gameboy(core_services *services, gb_video_renderer::render_callback &&r
         : _videoRenderer{services->videoRenderer(), renderCallback, 100}, _audioRenderer{services->audioRenderer()}, _gamepadSource{services->gamepadSource()}, _gb{&_videoRenderer, &_audioRenderer, &_gamepadSource, [this] {}, [link_cable_source] { return link_cable_source->readByte(); }, [link_cable_source](uint8_t data) { link_cable_source->sendByte(data); }}, _address_scanner{_gb.create_address_scanner()} {
 }
 
-void gameboy::load_rom(uint8_t *romData, uint32_t romLength, uint8_t *ram, uint32_t ramLength) {
+void gameboy::load_rom(uint8_t *romData, size_t romLength, uint8_t *ram, size_t ramLength) {
     _gb.load_rom(romData, romLength, ram, ramLength);
 }
